removing users, also refactored a bit
